chore: push용 커밋

+++ backend/src/main/java/com/ssafy/security/handler/CustomAuthenticationSuccessHandler.java
@@ -34,7 +34,7 @@ public class CustomAuthenticationSuccessHandler implements AuthenticationSuccess
     private final RefreshTokenService refreshTokenService;
 
     @Value("${jwt.refresh-expired}")
-    private Long refreshTokenExpired;
+    private int refreshTokenExpired;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
@@ -44,6 +44,7 @@ public class CustomAuthenticationSuccessHandler implements AuthenticationSuccess
         // 토큰 생성
         JwtDto jwtDto = jwtUtil.generateJwtDto(authentication);
 
+        // redis에 refreshToken 저장
         refreshTokenService.addRefreshToken(Long.valueOf(authentication.getName()), jwtDto.getRefreshToken());
 
         JwtResponse jwtResponse = JwtResponse.builder()
@@ -51,13 +52,15 @@ public class CustomAuthenticationSuccessHandler implements AuthenticationSuccess
                 .accessToken(jwtDto.getAccessToken())
                 .build();
 
+        // 응답객체 타입 설정
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+
+        // 클라이언트에게 응답 - 응답객체와 상태코드 추후 수정할것
         try(OutputStream os = response.getOutputStream()) {
             ObjectMapper objectMapper = new ObjectMapper();
             objectMapper.writeValue(os, ResponseEntity.ok(jwtResponse));
             os.flush();
         }
-
 
     }
